feat: implement user authentication and authorization
- add signup and login functionality
- implement session-based authentication
- restrict product creation to authenticated users
- restrict product creation to ADMIN role only
- display login status in UI

+++ src/main/java/com/filltex/price_table/domain/Product.java
@@ -30,12 +30,13 @@ public class Product {
 
     @Column(nullable = false)
     /** 제품 가격 */
-    private Integer price;
+    private Double price;
 
     /**
      * 제품 정보 수정
      */
-    public void update(String itemName, String finish, Integer price) {
+    public void update(String itemName, String finish, Double price) {
+
         this.itemName = itemName;
         this.finish = finish;
         this.price = price;
